Moved withregistry to a dedicated file and renamed pass_cls to with_subclasses :microscope:
